+++ apps/E2E/src/common/consts.ts
@@ -4,6 +4,7 @@ export const ROOT_VIEW = 'Fluent_Tester_Root_View';
 export const BUTTON_A11Y_ROLE = 'ControlType.Button';
 export const CALLOUT_A11Y_ROLE = 'ControlType.Group';
 export const CHECKBOX_A11Y_ROLE = 'ControlType.CheckBox';
+export const IMAGE_A11Y_ROLE = 'ControlType.Image';
 export const LINK_A11Y_ROLE = 'ControlType.HyperLink';
 export const MENUBUTTON_A11Y_ROLE = 'ControlType.Button';
 export const MENU_A11Y_ROLE = 'ControlType.Menu';
@@ -19,6 +20,7 @@ export const BOOT_APP_TIMEOUT = 60000;
 export const PAGE_TIMEOUT = 15000;
 
 export const enum Attribute {
+  AccessibilityHint = 'HelpText',
   AccessibilityLabel = 'Name',
   AccessibilityRole = 'ControlType',
   ExpandCollapseState = 'ExpandCollapse.ExpandCollapseState',
@@ -31,6 +33,7 @@ export const enum Attribute {
 }
 
 export const attributeToEnumName = {
+  [Attribute.AccessibilityHint]: 'AccessibilityHint',
   [Attribute.AccessibilityLabel]: 'AccessibilityLabel',
   [Attribute.AccessibilityRole]: 'AccessibilityRole',
   [Attribute.ExpandCollapseState]: 'ExpandCollapseState',
@@ -114,3 +117,10 @@ export const enum Keys {
   F11 = '\uE03B',
   F12 = '\uE03C',
 }
+
+// Used for E2E testing on base page of FluentTester
+export const BASE_TESTPAGE = 'Base_TestPage';
+
+// ScrollViewer that contains the list of test page buttons
+export const TESTPAGE_BUTTONS_SCROLLVIEWER = 'TestPage_Buttons_ScrollViewer';
+