+++ apps/fluent-tester/src/TestComponents/Avatar/E2EAvatarTest.tsx
@@ -8,7 +8,7 @@ import {
   AVATAR_ACCESSIBILITY_ROLE,
   AVATAR_TEST_COMPONENT,
   AVATAR_SECONDARY_TEST_COMPONENT,
-} from './consts';
+} from '../../../../E2E/src/Avatar/consts';
 
 export const E2EAvatarTest: React.FunctionComponent = () => {
   return (