+++ apps/fluent-tester/src/TestComponents/FocusZone/FocusZoneE2ETest.tsx
@@ -15,7 +15,7 @@ import {
   FOCUSZONE_GRID_AFTER,
   FOCUSZONE_GRID_BUTTON,
   FOCUSZONE_DIRECTION_PICKER,
-} from './consts';
+} from '../../../../E2E/src/FocusZone/consts';
 import { focusZoneTestStyles, GridButton } from './styles';
 import { testProps } from '../Common/TestProps';