+++ apps/fluent-tester/src/TestComponents/LinkExperimental/LinkTest.tsx
@@ -3,10 +3,11 @@ import { Alert, View, StyleSheet, Text, TextInput } from 'react-native';
 import { Link, LinkTokens } from '@fluentui-react-native/experimental-link';
 import { Stack } from '@fluentui-react-native/stack';
 import { stackStyle, commonTestStyles as commonStyles } from '../Common/styles';
-import { EXPERIMENTAL_LINK_TESTPAGE } from './consts';
+import { EXPERIMENTAL_LINK_TESTPAGE } from '../../../../E2E/src/LinkExperimental/consts';
 import { Test, TestSection, PlatformStatus } from '../Test';
 import { LinkE2ETest } from './E2ELinkTest';
 import { Platform } from 'react-native';
+import { InlineLinks } from './InlineLinksTest';
 
 const DefaultLinks: React.FunctionComponent = () => {
   const doPress = React.useCallback(() => Alert.alert('Alert.', 'You have been alerted.'), []);
@@ -25,52 +26,24 @@ const DefaultLinks: React.FunctionComponent = () => {
   );
 };
 
-const InlineLinks: React.FunctionComponent = () => {
-  const doPress = React.useCallback(() => Alert.alert('Alert.', 'You have been alerted.'), []);
-  const doAllyTap = React.useCallback(() => Alert.alert('Alert.', 'You have invoked onAllyTap.'), []);
-
-  return (
-    <Stack style={stackStyle}>
-      <Text>
-        Click{' '}
-        <Link inline onPress={doPress} onAccessibilityTap={doAllyTap}>
-          this link
-        </Link>{' '}
-        to alert me.
-      </Text>
-      <Text>
-        This{' '}
-        <Link inline onPress={doPress} disabled focusable>
-          link
-        </Link>{' '}
-        is disabled but focusable.
-      </Text>
-      <Text>
-        Follow this{' '}
-        <Link inline url="https://www.bing.com/">
-          link
-        </Link>{' '}
-        to navigate.
-      </Text>
-    </Stack>
-  );
-};
-
 const SubtleLinks: React.FunctionComponent = () => {
   const doPress = React.useCallback(() => Alert.alert('Alert.', 'You have been alerted.'), []);
   const doAllyTap = React.useCallback(() => Alert.alert('Alert.', 'You have invoked onAllyTap.'), []);
+  const supportsInlineLink = Platform.OS !== ('win32' as any);
 
   return (
     <Stack style={stackStyle}>
       <Link appearance="subtle" url="https://www.bing.com/">
         Click to navigate.
       </Link>
-      <Text>
-        This is inline Link.{' '}
-        <Link appearance="subtle" inline onPress={doPress} onAccessibilityTap={doAllyTap}>
-          Click to alert.
-        </Link>
-      </Text>
+      {supportsInlineLink && (
+        <Text>
+          This is inline Link.{' '}
+          <Link appearance="subtle" inline onPress={doPress} onAccessibilityTap={doAllyTap}>
+            Click to alert.
+          </Link>
+        </Text>
+      )}
       <Link appearance="subtle" onPress={doPress} disabled>
         Disabled Link
       </Link>
@@ -179,20 +152,22 @@ const linkSections: TestSection[] = [
     name: 'Inline Links',
     component: InlineLinks,
   },
-  ...Platform.select({
-    // As per design discussion , There is no use case for subtle link on Android , No tokens available for same.
-    android: [null],
-    default: [
-      {
-        name: 'Subtle Links',
-        component: SubtleLinks,
-      },
-    ],
+  Platform.select({
+    // As per design discussion, there is no use case for subtle link on Android, no tokens available for the same.
+    android: null,
+    default: {
+      name: 'Subtle Links',
+      component: SubtleLinks,
+    },
   }),
-  {
-    name: 'Custom Link',
-    component: CustomLinks,
-  },
+  Platform.select({
+    android: null,
+    default: {
+      name: 'Custom Link',
+      component: CustomLinks,
+    },
+  }),
+
   {
     name: 'Link E2E Test',
     component: LinkE2ETest,