+++ apps/fluent-tester/src/TestComponents/Menu/index.ts
@@ -1,2 +1,1 @@
 export * from './MenuTest';
-export * from './consts';