+++ packages/experimental/Link/src/Link.types.ts
@@ -7,6 +7,7 @@ export const linkName = 'Link';
 export type LinkState = IPressableState & {
   /**
    * Specifies whether the link has been visited.
+   * Note: Not supported for Android
    * @default false
    */
   visited?: boolean;
@@ -17,11 +18,14 @@ export type LinkState = IPressableState & {
   disabled?: boolean;
   /**
    * Specifies whether the link is inline.
+   * Note: Not supported for win32
+   *
    * @default false
    */
   inline?: boolean;
   /**
    * Specifies whether the link is subtle.
+   * Note: Not supported for Android
    * @default false
    */
   subtle?: boolean;
@@ -30,6 +34,7 @@ export type LinkState = IPressableState & {
 /**
  * Link tokens, these are the internally configurable values for Link elements. In particular these
  * drive decisions on how to build the styles
+ * Note: 'hovered','focused','visited','subtle' are not supported for Android
  */
 export interface LinkTokens extends TextTokens {
   hovered?: LinkTokens;
@@ -48,7 +53,9 @@ export type LinkAppearance = 'default' | 'subtle';
  */
 export interface LinkProps extends IWithPressableOptions<TextProps> {
   /**
-   * link appearance
+   * The appearance of the link, either `default` or `subtle`
+   * Note: 'subtle' is not supported for Android
+   * @default default
    */
   appearance?: LinkAppearance;
   /**
@@ -61,7 +68,9 @@ export interface LinkProps extends IWithPressableOptions<TextProps> {
    */
   enableFocusRing?: boolean;
   /**
-   * link inline
+   * Whether the link is inline with text
+   * Note: Not supported for win32
+   * @default false
    */
   inline?: boolean;
   /**
@@ -71,6 +80,7 @@ export interface LinkProps extends IWithPressableOptions<TextProps> {
   url?: string;
   /**
    * Text that should show in a tooltip when the user hovers over a button.
+   * Note: Not supported for Android
    */
   tooltip?: string;
 }