+++ packages/experimental/MenuButton/CHANGELOG.md
@@ -1,8 +1,84 @@
 # Change Log - @fluentui-react-native/experimental-menu-button
 
-This log was last generated on Thu, 15 Dec 2022 14:50:50 GMT and should not be manually modified.
+This log was last generated on Fri, 30 Dec 2022 16:09:27 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.12
+
+Fri, 30 Dec 2022 16:09:27 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.11
+- Bump @fluentui-react-native/experimental-button to v0.16.73
+- Bump @fluentui-react-native/framework to v0.8.30
+
+## 0.6.11
+
+Fri, 30 Dec 2022 03:49:05 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.10
+- Bump @fluentui-react-native/experimental-button to v0.16.72
+
+## 0.6.10
+
+Thu, 29 Dec 2022 08:02:57 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.9
+- Bump @fluentui-react-native/experimental-button to v0.16.71
+
+## 0.6.9
+
+Thu, 29 Dec 2022 03:47:47 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.8
+- Bump @fluentui-react-native/experimental-button to v0.16.70
+- Bump @fluentui-react-native/framework to v0.8.29
+- Bump @fluentui-react-native/tokens to v0.20.2
+
+## 0.6.8
+
+Wed, 28 Dec 2022 17:46:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-button to v0.16.69
+
+## 0.6.7
+
+Tue, 27 Dec 2022 22:21:15 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.7
+- Bump @fluentui-react-native/experimental-button to v0.16.68
+- Bump @fluentui-react-native/framework to v0.8.28
+- Bump @fluentui-react-native/tokens to v0.20.1
+
+## 0.6.6
+
+Tue, 27 Dec 2022 10:18:14 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.6
+- Bump @fluentui-react-native/experimental-button to v0.16.67
+- Bump @fluentui-react-native/framework to v0.8.27
+
+## 0.6.5
+
+Thu, 22 Dec 2022 19:31:49 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.5
 
 ## 0.6.4